Adding support for triggers requested for the ISR analysis.
In SusyNtuple, the trigger flags have been changed from uint to long long,
in order to have more than 32 triggers. The trigger flags in SusyD3PDAna
have been updated to long long.

I've added the requested trigger chains to the event flags in
SusyD3PDAna::fillEventTriggers, but have not yet added the object matching.

+++ SusyCommon/SusyD3PDAna.h
@@ -301,13 +301,13 @@ class SusyD3PDAna : public SusyD3PDInterface
     std::vector<int>            m_truJets;      // selected truth jets
     TLorentzVector              m_truMet;       // Truth MET
 
-    uint                        m_evtTrigFlags; // Event trigger flags
+    long long                   m_evtTrigFlags; // Event trigger flags
     
     // Trigger object matching maps
     // Key: d3pd index, Val: trig bit word
-    std::map<int, uint>         m_eleTrigFlags; // electron trigger matching flags
-    std::map<int, uint>         m_muoTrigFlags; // muon trigger matching flags
-    std::map<int, uint>         m_tauTrigFlags; // tau trigger matching flags
+    std::map<int, long long>    m_eleTrigFlags; // electron trigger matching flags
+    std::map<int, long long>    m_muoTrigFlags; // muon trigger matching flags
+    std::map<int, long long>    m_tauTrigFlags; // tau trigger matching flags
     
     float                       m_lumi;         // normalized luminosity (defaults to 4.7/fb)
     float                       m_sumw;         // sum of mc weights for normalization, must be set by user
